+++ src/hooks/queries/useHolidayListQuery.ts
@@ -1,31 +1,23 @@
 import { useQuery } from 'react-query';
-import { useRecoilValue, useSetRecoilState } from 'recoil';
+import { useRecoilValue } from 'recoil';
 import { ScheduleApi } from '@api/schedule';
 import { currentTimeAtom } from '@store/currentTime';
-import { holidayAtom } from '@store/holiday';
 
 const useHolidayListQuery = () => {
   const currentTime = useRecoilValue(currentTimeAtom);
-  const setHoliday = useSetRecoilState(holidayAtom);
 
   const useGetHolidayList = () => {
-    const { isLoading } = useQuery(
+    const { isLoading, data } = useQuery(
       ['getHoliday', currentTime.year()],
       async () => {
         const response = ScheduleApi.getHolidayList(`year=${currentTime.year()}`);
         return response;
       },
-      {
-        onSuccess: (data) => {
-          if (!data) {
-            return;
-          }
-          setHoliday(data.data);
-        },
-      }
+      {}
     );
     return {
       isLoading,
+      holidayList: data?.data || [],
     };
   };
