+++ src/hooks/useEventSchedule.ts
@@ -3,9 +3,7 @@ import { useTheme } from '@emotion/react';
 import moment from 'moment';
 import { useRouter } from 'next/router';
 import { useTranslation } from 'next-i18next';
-import { useRecoilValue } from 'recoil';
 import { DATE_FORMAT } from '@constants/format';
-import { currentMonthEventSelector } from '@store/eventSchedule';
 
 export enum EventPaintEnum {
   StartDate = 'startDate',
@@ -13,10 +11,6 @@ export enum EventPaintEnum {
   Ing = 'ing',
   OneDay = 'oneDay',
   Empty = 'empty',
-}
-
-interface TopPosition {
-  position: number;
 }
 
 export interface EventScheduleType {
@@ -27,13 +21,9 @@ export interface EventScheduleType {
   bgColor: string;
 }
 
-export type CurrentMonthEventType = EventScheduleType & TopPosition;
 type CbFunctionType = () => void;
 
 interface UseEventScheduleType {
-  currentMonthEvent: CurrentMonthEventType[];
-  currentDateMainlyEvent: EventScheduleType[];
-  currentDateHolidayEvent: EventScheduleType[];
   getEventPaintType: (event: EventScheduleType, date: string) => EventPaintEnum;
   handleClickDate: (date: string, callback: CbFunctionType) => void;
   handleClickSchedule: (event: EventScheduleType, callback: CbFunctionType) => void;
@@ -42,23 +32,9 @@ interface UseEventScheduleType {
 }
 
 const useEventSchedule = (): UseEventScheduleType => {
-  const currentMonthEvent = useRecoilValue(currentMonthEventSelector);
-
   const { query, replace } = useRouter();
   const { i18n } = useTranslation();
   const { colors } = useTheme();
-
-  const currentDateMainlyEvent = useMemo(() => {
-    const { date } = query;
-    if (!date) return [];
-    return currentMonthEvent.filter((event) => event.startDate <= date && event.endDate >= date && event.typeId !== 2);
-  }, [query, currentMonthEvent]);
-
-  const currentDateHolidayEvent = useMemo(() => {
-    const { date } = query;
-    if (!date) return [];
-    return currentMonthEvent.filter((event) => event.startDate <= date && event.endDate >= date && event.typeId === 2);
-  }, [query, currentMonthEvent]);
 
   const boardDateTitle = useMemo(() => {
     return moment(query?.date).format(
@@ -122,9 +98,6 @@ const useEventSchedule = (): UseEventScheduleType => {
   }, [query]);
 
   return {
-    currentMonthEvent,
-    currentDateMainlyEvent,
-    currentDateHolidayEvent,
     boardDateTitle,
     getEventPaintType,
     handleClickDate,
