--- src/store/holiday.ts
@@ -1,7 +1,0 @@
-import { atom } from 'recoil';
-import { EventScheduleType } from '@hooks/useEventSchedule';
-
-export const holidayAtom = atom<EventScheduleType[]>({
-  key: 'holidayState',
-  default: [],
-});
